Add component name & slug validation Improve code climate

+++ src/Core/Tests/Domain/CollectionTest.php
@@ -36,6 +36,7 @@
 
 namespace Artefakt\Core\Tests\Domain;
 
+use Artefakt\Core\Domain\Contract\AbstractNodeInterface;
 use Artefakt\Core\Domain\Model\Collection;
 use Artefakt\Core\Domain\Model\Component;
 use Artefakt\Core\Tests\AbstractTestBase;
@@ -78,7 +79,8 @@ class CollectionTest extends AbstractTestBase
             $this->assertEquals(array_shift($components), $component);
         }
 
-        $this->assertEquals(1, count($collection->detach($component)));
+        if ($component instanceof AbstractNodeInterface) {
+            $this->assertEquals(1, count($collection->detach($component)));
+        }
     }
-
 }
